Move state-options.ts to utils folder - update import paths
